feat(settings): T2 restructure — tabs + URL-hash + search + sticky save (#399)
Split the 2,328-line SettingsView.jsx into a parent + 9 per-tab
subcomponents under web/js/components/preact/settings/. Parent retains
shared state, save wiring, viewer-mode gating, restart notice; each tab
is a pure-JSX subcomponent accepting settings + handlers via props.

- URL-hash persistence (#general, #go2rtc, ...) via history.replaceState,
  listens to hashchange for browser nav, defaults to 'general'.
- Live label search (150 ms debounce) walks [data-setting-label] across
  all mounted tab panels, highlights matches, dims non-matches,
  auto-switches to the tab owning the first match, smooth-scrolls into
  view. Clear (×) button included.
- Sticky save bar with T1's <AsyncButton> wiring preserved verbatim.
  Disabled until settings differ from the baseline snapshot. Safe-area
  padding for iOS gesture bar.
- Mobile chrome: horizontal snap-mandatory tab strip at top (< 640 px).
  Desktop: vertical sidebar pill list (≥ 640 px). Mirrors SystemView's
  tab pattern and honors the T6 44-px touch-target floor.
- Preserved T4 Reduce Motion toggle verbatim in AppearanceTab.
- Preserved T9/T14/T4b go2rtc override editor verbatim in Go2rtcTab
  (quarantine banner, size indicator, validate-on-blur, effective-config
  modal, presets dropdown, supported-sections collapsible).
- Preserved viewer-mode short-circuit: viewers see only AppearanceTab +
  admin-only notice; no tab chrome, no search, no save.
- 15 new locale keys added to en.json + pt-BR.json (tab labels, search
  placeholder, "Unsaved changes", Authentication & Access header,
  Streams Defaults header + description). Other locales fall back via
  the existing t(key) || key path.

Verification:
- npm --prefix web run build → ✓ built in 10.24s, 0 errors.
- npm --prefix web test -- tests/useAsyncAction.spec.js → 4/4 pass.
- Full Jest suite: only the pre-existing i18n "handles complete locale
  loading failure gracefully" failure remains (flagged by T1 and T5).
- 71 onChange={handleInputChange} callsites before → 71 after, every
  {settings.X} state access moved to exactly one tab.
