+++ web/public/locales/en.json
@@ -136,6 +136,21 @@
   "settings.restartNotice": "Restart LightNVR before runtime capacity changes take effect.",
   "settings.restartExplanation": "Until LightNVR restarts, the running instance keeps using the old stream capacity and thread pool.",
   "settings.save": "Save Settings",
+  "settings.tab.general": "General",
+  "settings.tab.storage": "Storage",
+  "settings.tab.streamsDefaults": "Streams Defaults",
+  "settings.tab.detection": "Detection",
+  "settings.tab.go2rtc": "go2rtc",
+  "settings.tab.mqtt": "MQTT",
+  "settings.tab.auth": "Auth / Security",
+  "settings.tab.appearance": "Appearance",
+  "settings.tab.advanced": "Advanced",
+  "settings.tabsLabel": "Settings sections",
+  "settings.searchPlaceholder": "Search settings…",
+  "settings.unsavedChanges": "Unsaved changes",
+  "settings.authentication": "Authentication & Access",
+  "settings.streamsDefaults": "Streams Defaults",
+  "settings.streamsDefaultsDescription": "Defaults applied to new streams. Per-stream overrides always win.",
   "users.userManagement": "User Management",
   "users.addNewUser": "Add New User",
   "users.addUser": "Add User",
