+++ web/public/locales/pt-BR.json
@@ -114,6 +114,21 @@
   "settings.restartNotice": "Reinicie o LightNVR antes que as mudanças de capacidade em tempo de execução entrem em vigor.",
   "settings.restartExplanation": "Até o LightNVR reiniciar, a instância em execução continuará usando a capacidade antiga de streams e do pool de threads.",
   "settings.save": "Salvar configurações",
+  "settings.tab.general": "Geral",
+  "settings.tab.storage": "Armazenamento",
+  "settings.tab.streamsDefaults": "Padrões de streams",
+  "settings.tab.detection": "Detecção",
+  "settings.tab.go2rtc": "go2rtc",
+  "settings.tab.mqtt": "MQTT",
+  "settings.tab.auth": "Autenticação / Segurança",
+  "settings.tab.appearance": "Aparência",
+  "settings.tab.advanced": "Avançado",
+  "settings.tabsLabel": "Seções de configurações",
+  "settings.searchPlaceholder": "Pesquisar configurações…",
+  "settings.unsavedChanges": "Alterações não salvas",
+  "settings.authentication": "Autenticação e acesso",
+  "settings.streamsDefaults": "Padrões de streams",
+  "settings.streamsDefaultsDescription": "Padrões aplicados a novos streams. Sobrescritas por stream sempre prevalecem.",
   "users.userManagement": "Gerenciamento de usuários",
   "users.addNewUser": "Adicionar novo usuário",
   "users.addUser": "Adicionar usuário",
